ピクセルデータの取り出しに成功
