--- src/photo/Main.java
@@ -1,16 +1,0 @@
-package photo;
-
-import org.opencv.core.Core;
-import org.opencv.core.CvType;
-import org.opencv.core.Mat;
-
-public class Main {
-
-    public static void main(String[] args) {
-        System.out.println("Welcome to OpenCV " + Core.VERSION);
-        System.loadLibrary(Core.NATIVE_LIBRARY_NAME);
-        Mat m  = Mat.eye(3, 3, CvType.CV_8UC1);
-        System.out.println("m = " + m.dump());
-    }
-
-}
